+++ CHANGELOG.md
@@ -15,6 +15,7 @@
   that [Trakt will only store and return minute-precision timestamps](https://github.com/trakt/trakt-api/discussions/694).
 * Add `TraktV2.isAccountLimitExceeded()`, `TraktV2.isRateLimitExceeded()` and `TraktV2.isServerError()` methods to help
   inspect `Response` objects for more common Trakt HTTP status codes.
+* Add variants of the watchlist methods that support pagination and sort order parameters.
 
 ## 6.16.0 - 2024-11-07